refactor: reduce bulk job response duplication
